apps/server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Piyush Jaiswal <iampiyushjaiswal103@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
base58 = "^2.1.1"
pynacl = "^1.5.0"
solana = "^0.32.0"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
